formatting

+++ src/utils/measurement.py
@@ -1,16 +1,17 @@
-from datetime import datetime
-import os
-import pandas as pd 
-import openpyxl
-import xml.etree.ElementTree as ET
-from config.database import Mdb
-from pathlib import Path
 import locale
-from utils import open_excel_file
+import os
+import xml.etree.ElementTree as ET
 
-locale.setlocale(locale.LC_TIME, 'Portuguese_Brazil.1252') # Necessário para colocar a data
+import pandas as pd
+
+from config.database import Mdb
+
+locale.setlocale(
+    locale.LC_TIME, "Portuguese_Brazil.1252"
+)  # Necessário para colocar a data
 
 ROUGH_DATA_FOLDER = r"C:/HPI_BWF/rough_data/"
+
 
 def get_rough_data(rough_data_locations: pd.Series):
     locations = rough_data_locations.unique()
@@ -42,33 +43,44 @@ def get_rough_data(rough_data_locations: pd.Series):
             print(f"Error parsing file {location}: {e}")
     return pd.DataFrame(data)
 
+
 def raw_extracted_data():
     with Mdb() as db:
-        dataframe = db.fetch('measurements')
+        dataframe = db.fetch("measurements")
 
-    rough_data = get_rough_data(dataframe['data_file'])
-    dataframe = dataframe.merge(rough_data, how='inner', on=['data_file', 'position_id'])
+    rough_data = get_rough_data(dataframe["data_file"])
+    dataframe = dataframe.merge(
+        rough_data, how="inner", on=["data_file", "position_id"]
+    )
 
     return dataframe
 
+
 def transform_columns(df: pd.DataFrame):
-    df['diameter_land'] = df['diameter_land'].astype(float) 
-    df['diameter_groove'] = df['diameter_groove'].astype(float)
-    df['measurement_date'] = df['measurement_date'].dt.date
+    df["diameter_land"] = df["diameter_land"].astype(float)
+    df["diameter_groove"] = df["diameter_groove"].astype(float)
+    df["measurement_date"] = df["measurement_date"].dt.date
 
     return df
+
 
 def extract_measurement_data(archives: list = []):
     dataframe = raw_extracted_data()
 
     if archives:
-        dataframe = dataframe[dataframe['archive_name'].isin(archives)]
+        dataframe = dataframe[dataframe["archive_name"].isin(archives)]
 
     dataframe = transform_columns(dataframe)
-    columns = ['measurement_date', 'measurement_name', 'depth', 'diameter_land', 'diameter_groove']
-    
-    if dataframe['diameter_groove'].max() == 0:
-        columns.remove('diameter_groove')
+    columns = [
+        "measurement_date",
+        "measurement_name",
+        "depth",
+        "diameter_land",
+        "diameter_groove",
+    ]
+
+    if dataframe["diameter_groove"].max() == 0:
+        columns.remove("diameter_groove")
 
     dataframe = dataframe[columns]
